requirements.txt
fastapi
uvicorn
kubernetes
pydantic
spacy
python-dotenv
openai==0.28
Flask
requests